fixed rate-players bug by separating page into client and server components. also avoided an error by ensuring groupid loads in authutils
